+++ lib/ui/athletics/AthleticsNewsArticlePanel.dart
@@ -196,9 +196,11 @@ class _AthleticsNewsArticlePanelState extends State<AthleticsNewsArticlePanel> w
   ],);
   }
 
+  static const String _shareReasonMacro = '{{reason}}';
+
   Future<void> _onShareArticle() async {
     Analytics().logSelect(target: "Share Article");
-    String? message;
+    String? message, reason;
     if (_articleUri?.isValid == true) {
       try {
         ShareResult result = await SharePlus.instance.share(ShareParams(uri: _articleUri!));
@@ -208,6 +210,8 @@ class _AthleticsNewsArticlePanelState extends State<AthleticsNewsArticlePanel> w
       }
       catch (e) {
         message = Localization().getStringEx('panel.athletics_news_article.message.share.failed.text', 'Failed to share article.');
+        reason = Localization().getStringEx('panel.athletics_news_article.message.share.reason.format', 'Reason: $_shareReasonMacro').
+          replaceAll(_shareReasonMacro, e.toString());
       }
     }
     else {
@@ -215,7 +219,13 @@ class _AthleticsNewsArticlePanelState extends State<AthleticsNewsArticlePanel> w
     }
 
     if (mounted && (message != null)) {
-      AppAlert.showTextMessage(context, message);
+      Widget messageText = (reason != null) ? Column(mainAxisSize: MainAxisSize.min, children: [
+        Text(message, textAlign: TextAlign.center, style: Styles().textStyles.getTextStyle('widget.message.large'),),
+        Padding(padding: EdgeInsets.only(top: 24), child:
+          Text(reason, textAlign: TextAlign.left, style: Styles().textStyles.getTextStyle('widget.message.medium.thin'),),
+        ),
+      ],) : Text(message, textAlign: TextAlign.center, style: Styles().textStyles.getTextStyle('widget.message.medium.thin'),);
+      AppAlert.showWidgetMessage(context, messageText, buttonTextStyle: Styles().textStyles.getTextStyle('widget.message.medium.thin'));
     }
   }
 
@@ -226,7 +236,7 @@ class _AthleticsNewsArticlePanelState extends State<AthleticsNewsArticlePanel> w
         padding: EdgeInsets.only(bottom: 10),
         child:
         HtmlWidget(
-            StringUtils.ensureNotEmpty(_article!.description),
+          StringUtils.ensureNotEmpty(_article!.description),
           textStyle:  Styles().textStyles.getTextStyle("widget.item.regular.thin"),
         )
       ));